refactor : 리스트 페이지 폴더 구조 재구
